chore: simplify file field state logic
Also:
- Alias the ImageField to FileField
- Add hook for useDraftDocValue() which has a return value similar to
  useState() but automatically subscribes to changes to the db
