+++ packages/root-cms/ui/components/DocEditor/fields/FileField.css
@@ -1,37 +1,37 @@
-.FileUploadField {
+.FileField {
   border: 1px solid #ced4da;
   padding: 10px;
   position: relative;
 }
 
-.FileUploadField:has(.dragging) {
+.FileField:has(.dragging) {
   border: 1px solid #eebefa;
 }
 
-.FileUploadField:has(.dragging) .FileUploadField__Dropzone,
-.FileUploadField:has(.dragging) .FileUploadField__Canvas {
+.FileField:has(.dragging) .FileField__Dropzone,
+.FileField:has(.dragging) .FileField__Canvas {
   background-color: #f3d9fa;
 }
 
-.FileUploadField:has(.FileUploadField__Canvas) {
+.FileField:has(.FileField__Canvas) {
   padding: 0;
 }
 
-.FileUploadField__Empty {
+.FileField__Empty {
   align-items: center;
   display: flex;
   flex-direction: column;
   gap: 4px;
 }
 
-.FileUploadField__Preview {
+.FileField__Preview {
   padding: 10px;
   pointer-events: none;
   position: relative;
   z-index: 1;
 }
 
-.FileUploadField__Preview__InfoButton {
+.FileField__Preview__InfoButton {
   pointer-events: all;
   position: absolute;
   top: 8px;
@@ -42,16 +42,16 @@
   gap: 4px;
 }
 
-.FileUploadField__Preview__InfoButton__Icon {
+.FileField__Preview__InfoButton__Icon {
   color: #333;
   background-color: white;
 }
 
-.FileUploadField__Preview__InfoButton__Icon:hover {
+.FileField__Preview__InfoButton__Icon:hover {
   background-color: #efefef;
 }
 
-.FileUploadField__reupload {
+.FileField__reupload {
   pointer-events: all;
   position: absolute;
   bottom: 8px;
@@ -59,11 +59,11 @@
   z-index: 1;
 }
 
-.FileUploadField__Canvas--infoOpened .FileUploadField__reupload {
+.FileField__Canvas--infoOpened .FileField__reupload {
   display: none;
 }
 
-.FileUploadField__Preview:has(.FileUploadField__Canvas) {
+.FileField__Preview:has(.FileField__Canvas) {
   padding: 0;
 }
 
@@ -92,7 +92,7 @@
   z-index: 2;
 }
 
-.FileUploadField__Preview__Info {
+.FileField__Preview__Info {
   position: absolute;
   top: 0px;
   left: 0px;
@@ -105,11 +105,11 @@
   z-index: 1;
 }
 
-.FileUploadField__Preview__Actions__Trash {
+.FileField__Preview__Actions__Trash {
   pointer-events: all;
 }
 
-.FileUploadField__Canvas__Info {
+.FileField__Canvas__Info {
   pointer-events: all;
   background-color: #f5f5f5;
   padding: 40px 10px 10px;
@@ -117,27 +117,27 @@
   overflow: auto;
 }
 
-.FileUploadField__Canvas__Info__Icon {
+.FileField__Canvas__Info__Icon {
   position: absolute;
   top: 12px;
   left: 10px;
 }
 
-.FileUploadField__Canvas__InfoTable {
+.FileField__Canvas__InfoTable {
   border-radius: 2px;
   background-color: white;
 }
 
-.FileUploadField__Canvas__InfoTable td {
+.FileField__Canvas__InfoTable td {
   padding: 5px;
   vertical-align: top;
 }
 
-.FileUploadField__Canvas__InfoTable textarea {
+.FileField__Canvas__InfoTable textarea {
   padding: 5px;
 }
 
-.FileUploadField__Canvas {
+.FileField__Canvas {
   aspect-ratio: var(--canvas-aspect-ratio, 16 / 9);
   background-color: var(--canvas-bg-color, #fff);
   max-height: var(--canvas-max-height, 280px);
@@ -146,20 +146,20 @@
   width: 100%;
 }
 
-.FileUploadField__Canvas:has(video) {
+.FileField__Canvas:has(video) {
   background-color: var(--canvas-bg-color, #000);
 }
 
-.FileUploadField__Canvas--infoOpened {
+.FileField__Canvas--infoOpened {
   max-height: unset;
   aspect-ratio: unset;
 }
 
-.FileUploadField__Canvas.dragging {
+.FileField__Canvas.dragging {
   background-color: #f3d9fa;
 }
 
-.FileUploadField__Preview__Image {
+.FileField__Preview__Image {
   margin: 0 auto;
   width: var(--canvas-asset-width, 100%);
   max-width: 100%;
@@ -169,13 +169,13 @@
   pointer-events: all;
 }
 
-.FileUploadField__Preview__Actions {
+.FileField__Preview__Actions {
   display: flex;
   padding: 10px;
   justify-content: space-between;
 }
 
-.FileUploadField__Dropzone {
+.FileField__Dropzone {
   appearance: none;
   border: 0;
   background: #f5f5f5;
@@ -184,44 +184,44 @@
   inset: 0;
 }
 
-.FileUploadField__InvisibleDropzone {
+.FileField__InvisibleDropzone {
   display: contents;
 }
 
-.FileUploadField__Dropzone:focus {
+.FileField__Dropzone:focus {
   outline: 1px solid -webkit-focus-ring-color;
 }
 
-.FileUploadField__Dropzone.dragging {
+.FileField__Dropzone.dragging {
   background-color: #f3d9fa;
 }
 
-.FileUploadField__Empty:focus {
+.FileField__Empty:focus {
   outline: 1px solid -webkit-focus-ring-color;
 }
 
-.FileUploadField__Empty__Label {
+.FileField__Empty__Label {
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   gap: 8px;
 }
 
-.FileUploadField__FileUploadButton {
+.FileField__FileUploadButton {
   font-size: 12px;
 }
 
-.FileUploadField__FileUploadButton--compact {
+.FileField__FileUploadButton--compact {
   padding: 2px 6px;
   gap: 6px;
   font-size: 12px;
 }
 
-.FileUploadField__FileUploadButton--compact .mantine-Button-leftIcon {
+.FileField__FileUploadButton--compact .mantine-Button-leftIcon {
   margin-right: 6px;
 }
 
-.FileUploadField__Empty__AcceptTypes {
+.FileField__Empty__AcceptTypes {
   font-size: 10px;
   display: inline;
   position: relative;
