+++ packages/root-cms/ui/components/DocEditor/fields/StringField.tsx
@@ -1,24 +1,27 @@
 import {TextInput, Textarea} from '@mantine/core';
 import {ChangeEvent} from 'preact/compat';
-import {useState} from 'preact/hooks';
 import * as schema from '../../../../core/schema.js';
-import {useDraftDoc, useDraftDocField} from '../../../hooks/useDraftDoc.js';
+import {useDraftDocValue} from '../../../hooks/useDraftDoc.js';
 import {requestHighlightNode} from '../../../utils/iframe-preview.js';
 import {FieldProps} from './FieldProps.js';
 
 export function StringField(props: FieldProps) {
   const field = props.field as schema.StringField;
-  const [value, setValue] = useState('');
-  const draft = useDraftDoc().controller;
+  const [value, setValue] = useDraftDocValue(props.deepKey, '');
 
-  const onChange = (newValue: string) => {
-    setValue(newValue);
-    draft.updateKey(props.deepKey, newValue);
+  const onChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setValue(e.currentTarget.value);
   };
 
-  useDraftDocField(props.deepKey, (newValue: string) => {
-    setValue(newValue);
-  });
+  const onFocus = (e: FocusEvent) => {
+    props.onFocus?.(e);
+    requestHighlightNode(props.deepKey, {scroll: true});
+  };
+
+  const onBlur = (e: FocusEvent) => {
+    props.onBlur?.(e);
+    requestHighlightNode(null);
+  };
 
   if (field.variant === 'textarea') {
     return (
@@ -29,17 +32,9 @@ export function StringField(props: FieldProps) {
         minRows={4}
         maxRows={field.maxRows || 12}
         value={value}
-        onChange={(e: ChangeEvent<HTMLTextAreaElement>) => {
-          onChange(e.currentTarget.value);
-        }}
-        onFocus={(e: FocusEvent) => {
-          props.onFocus?.(e);
-          requestHighlightNode(props.deepKey, {scroll: true});
-        }}
-        onBlur={(e: FocusEvent) => {
-          props.onBlur?.(e);
-          requestHighlightNode(null);
-        }}
+        onChange={onChange}
+        onFocus={onFocus}
+        onBlur={onBlur}
       />
     );
   }
@@ -48,17 +43,9 @@ export function StringField(props: FieldProps) {
       size="xs"
       radius={0}
       value={value}
-      onChange={(e: ChangeEvent<HTMLInputElement>) => {
-        onChange(e.currentTarget.value);
-      }}
-      onFocus={(e: FocusEvent) => {
-        props.onFocus?.(e);
-        requestHighlightNode(props.deepKey, {scroll: true});
-      }}
-      onBlur={(e: FocusEvent) => {
-        props.onBlur?.(e);
-        requestHighlightNode(null);
-      }}
+      onChange={onChange}
+      onFocus={onFocus}
+      onBlur={onBlur}
     />
   );
 }
